Add Backfill support to internal statestore (#1273)
